pyproject.toml
[tool.poetry]
name = "engram3"
version = "0.1.0"
description = "Keyboard layout optimization through preference learning"
authors = ["Arno Klein <binarybottle@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"  # Changed from ^3.8 to >=3.9
pandas = "^1.5.0"
numpy = "^1.23.0"
pyyaml = "^6.0"
setuptools = "^69.0.2"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
